requirements.txt
Pillow==7.1.2
urllib3==1.24.2
